refactor(cli): move create to handler + helpers
