+++ quartz/cli/bootstrap-cli.mjs
@@ -17,7 +17,8 @@ import serveHandler from "serve-handler"
 import { WebSocketServer } from "ws"
 import { randomUUID } from "crypto"
 import { Mutex } from "async-mutex"
-import { handleBuild } from "./cli-functions.js"
+import { handleBuild } from "./cli-handlers.js"
+import { CommonArgv, BuildArgv, CreateArgv, SyncArgv } from "./args.js"
 
 const ORIGIN_NAME = "origin"
 const UPSTREAM_NAME = "upstream"
@@ -28,112 +29,6 @@ const cacheFile = "./.quartz-cache/transpiled-build.mjs"
 const fp = "./quartz/build.ts"
 const { version } = JSON.parse(readFileSync("./package.json").toString())
 const contentCacheFolder = path.join(cacheDir, "content-cache")
-
-const CommonArgv = {
-  directory: {
-    string: true,
-    alias: ["d"],
-    default: "content",
-    describe: "directory to look for content files",
-  },
-  verbose: {
-    boolean: true,
-    alias: ["v"],
-    default: false,
-    describe: "print out extra logging information",
-  },
-}
-
-const CreateArgv = {
-  ...CommonArgv,
-  source: {
-    string: true,
-    alias: ["s"],
-    describe: "source directory to copy/create symlink from",
-  },
-  strategy: {
-    string: true,
-    alias: ["X"],
-    choices: ["new", "copy", "symlink"],
-    describe: "strategy for content folder setup",
-  },
-  links: {
-    string: true,
-    alias: ["l"],
-    choices: ["absolute", "shortest", "relative"],
-    describe: "strategy to resolve links",
-  },
-}
-
-const SyncArgv = {
-  ...CommonArgv,
-  commit: {
-    boolean: true,
-    default: true,
-    describe: "create a git commit for your unsaved changes",
-  },
-  push: {
-    boolean: true,
-    default: true,
-    describe: "push updates to your Quartz fork",
-  },
-  pull: {
-    boolean: true,
-    default: true,
-    describe: "pull updates from your Quartz fork",
-  },
-}
-
-const BuildArgv = {
-  ...CommonArgv,
-  output: {
-    string: true,
-    alias: ["o"],
-    default: "public",
-    describe: "output folder for files",
-  },
-  serve: {
-    boolean: true,
-    default: false,
-    describe: "run a local server to live-preview your Quartz",
-  },
-  baseDir: {
-    string: true,
-    default: "",
-    describe: "base path to serve your local server on",
-  },
-  port: {
-    number: true,
-    default: 8080,
-    describe: "port to serve Quartz on",
-  },
-  bundleInfo: {
-    boolean: true,
-    default: false,
-    describe: "show detailed bundle information",
-  },
-  concurrency: {
-    number: true,
-    describe: "how many threads to use to parse notes",
-  },
-}
-
-function escapePath(fp) {
-  return fp
-    .replace(/\\ /g, " ") // unescape spaces
-    .replace(/^".*"$/, "$1")
-    .replace(/^'.*"$/, "$1")
-    .trim()
-}
-
-function exitIfCancel(val) {
-  if (isCancel(val)) {
-    outro(chalk.red("Exiting"))
-    process.exit(0)
-  } else {
-    return val
-  }
-}
 
 async function stashContentFolder(contentFolder) {
   await fs.promises.rm(contentCacheFolder, { force: true, recursive: true })
@@ -170,170 +65,7 @@ yargs(hideBin(process.argv))
   .version(version)
   .usage("$0 <cmd> [args]")
   .command("create", "Initialize Quartz", CreateArgv, async (argv) => {
-    console.log()
-    intro(chalk.bgGreen.black(` Quartz v${version} `))
-    const contentFolder = path.join(cwd, argv.directory)
-    let setupStrategy = argv.strategy?.toLowerCase()
-    let linkResolutionStrategy = argv.links?.toLowerCase()
-    const sourceDirectory = argv.source
-
-    // If all cmd arguments were provided, check if theyre valid
-    if (setupStrategy && linkResolutionStrategy) {
-      // If setup isn't, "new", source argument is required
-      if (setupStrategy !== "new") {
-        // Error handling
-        if (!sourceDirectory) {
-          outro(
-            chalk.red(
-              `Setup strategies (arg '${chalk.yellow(
-                `-${CreateArgv.strategy.alias[0]}`,
-              )}') other than '${chalk.yellow(
-                "new",
-              )}' require content folder argument ('${chalk.yellow(
-                `-${CreateArgv.source.alias[0]}`,
-              )}') to be set`,
-            ),
-          )
-          process.exit(1)
-        } else {
-          if (!fs.existsSync(sourceDirectory)) {
-            outro(
-              chalk.red(
-                `Input directory to copy/symlink 'content' from not found ('${chalk.yellow(
-                  sourceDirectory,
-                )}', invalid argument "${chalk.yellow(`-${CreateArgv.source.alias[0]}`)})`,
-              ),
-            )
-            process.exit(1)
-          } else if (!fs.lstatSync(sourceDirectory).isDirectory()) {
-            outro(
-              chalk.red(
-                `Source directory to copy/symlink 'content' from is not a directory (found file at '${chalk.yellow(
-                  sourceDirectory,
-                )}', invalid argument ${chalk.yellow(`-${CreateArgv.source.alias[0]}`)}")`,
-              ),
-            )
-            process.exit(1)
-          }
-        }
-      }
-    }
-
-    // Use cli process if cmd args werent provided
-    if (!setupStrategy) {
-      setupStrategy = exitIfCancel(
-        await select({
-          message: `Choose how to initialize the content in \`${contentFolder}\``,
-          options: [
-            { value: "new", label: "Empty Quartz" },
-            { value: "copy", label: "Copy an existing folder", hint: "overwrites `content`" },
-            {
-              value: "symlink",
-              label: "Symlink an existing folder",
-              hint: "don't select this unless you know what you are doing!",
-            },
-          ],
-        }),
-      )
-    }
-
-    async function rmContentFolder() {
-      const contentStat = await fs.promises.lstat(contentFolder)
-      if (contentStat.isSymbolicLink()) {
-        await fs.promises.unlink(contentFolder)
-      } else {
-        await rimraf(contentFolder)
-      }
-    }
-
-    await fs.promises.unlink(path.join(contentFolder, ".gitkeep"))
-    if (setupStrategy === "copy" || setupStrategy === "symlink") {
-      let originalFolder = sourceDirectory
-
-      // If input directory was not passed, use cli
-      if (!sourceDirectory) {
-        originalFolder = escapePath(
-          exitIfCancel(
-            await text({
-              message: "Enter the full path to existing content folder",
-              placeholder:
-                "On most terminal emulators, you can drag and drop a folder into the window and it will paste the full path",
-              validate(fp) {
-                const fullPath = escapePath(fp)
-                if (!fs.existsSync(fullPath)) {
-                  return "The given path doesn't exist"
-                } else if (!fs.lstatSync(fullPath).isDirectory()) {
-                  return "The given path is not a folder"
-                }
-              },
-            }),
-          ),
-        )
-      }
-
-      await rmContentFolder()
-      if (setupStrategy === "copy") {
-        await fs.promises.cp(originalFolder, contentFolder, {
-          recursive: true,
-          preserveTimestamps: true,
-        })
-      } else if (setupStrategy === "symlink") {
-        await fs.promises.symlink(originalFolder, contentFolder, "dir")
-      }
-    } else if (setupStrategy === "new") {
-      await fs.promises.writeFile(
-        path.join(contentFolder, "index.md"),
-        `---
-title: Welcome to Quartz
----
-
-This is a blank Quartz installation.
-See the [documentation](https://quartz.jzhao.xyz) for how to get started.
-`,
-      )
-    }
-
-    // Use cli process if cmd args werent provided
-    if (!linkResolutionStrategy) {
-      // get a preferred link resolution strategy
-      linkResolutionStrategy = exitIfCancel(
-        await select({
-          message: `Choose how Quartz should resolve links in your content. You can change this later in \`quartz.config.ts\`.`,
-          options: [
-            {
-              value: "absolute",
-              label: "Treat links as absolute path",
-              hint: "for content made for Quartz 3 and Hugo",
-            },
-            {
-              value: "shortest",
-              label: "Treat links as shortest path",
-              hint: "for most Obsidian vaults",
-            },
-            {
-              value: "relative",
-              label: "Treat links as relative paths",
-              hint: "for just normal Markdown files",
-            },
-          ],
-        }),
-      )
-    }
-
-    // now, do config changes
-    const configFilePath = path.join(cwd, "quartz.config.ts")
-    let configContent = await fs.promises.readFile(configFilePath, { encoding: "utf-8" })
-    configContent = configContent.replace(
-      /markdownLinkResolution: '(.+)'/,
-      `markdownLinkResolution: '${linkResolutionStrategy}'`,
-    )
-    await fs.promises.writeFile(configFilePath, configContent)
-
-    outro(`You're all set! Not sure what to do next? Try:
-   • Customizing Quartz a bit more by editing \`quartz.config.ts\`
-   • Running \`npx quartz build --serve\` to preview your Quartz locally
-   • Hosting your Quartz online (see: https://quartz.jzhao.xyz/hosting)
-`)
+    //
   })
   .command("update", "Get the latest Quartz updates", CommonArgv, async (argv) => {
     const contentFolder = path.join(cwd, argv.directory)
